daybed views refactoring
